improve checkout / branch completion
Improves the script itself (using flags for the directory / repository)
and the completions that make use of the script. For example, previously
when typing `w <repo> o`, only local branches that start with `o` have
been shown. With these improvements, using `commandline -ct`, we now
pass the `o` to `_c_complete` as well, which will then suggest remote
/ origin branches as well.

(`commandline -ct` *c*uts at the cursor and prints the current
*t*oken)